Change database from PostgreSQL to MS SQL Server.

+++ SaleSystemData/Models/Product.cs
@@ -13,7 +13,7 @@ namespace SaleSystemData.Models
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string Description { get; set; }
-        public decimal? Price { get; set; }
+        public double? Price { get; set; }
         public string Unit { get; set; }
 
         public virtual ICollection<Order> Order { get; set; }
